+++ src-tauri/src/lib.rs
@@ -51,6 +51,7 @@ pub fn run() {
   ];
 
   tauri::Builder::default()
+    .plugin(tauri_plugin_store::Builder::new().build())
     .plugin(tauri_plugin_fs::init())
     .plugin(
       tauri_plugin_sql::Builder::default()