+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 ### Added
 - Account opening balances during creation
-- Basic currency selection
+- Basic currency selection and default currency option
+- User settings persistence through Tauri's `plugin-store`
 
 ### Fixed
 - Negative zero liabilities in account summary